+++ src/reducers/composerReducer.js
@@ -1,11 +1,23 @@
-export default (
-  state = {
-    draftStatus: "Draft",
-    editorFocusRequested: 0
-  },
-  action
-) => {
+const INITIAL_STATE = {
+  draftStatus: "Draft",
+  editorFocusRequested: 0,
+  uploadProgress: 0
+}
+
+export default (state = INITIAL_STATE, action) => {
   switch (action.type) {
+    case "UPLOAD.SET_STATUS":
+      state = {
+        ...state,
+        uploadProgress: action.payload.uploadProgress
+      }
+      break
+    case "UPLOAD.INIT_STATUS":
+      state = {
+        ...state,
+        uploadProgress: INITIAL_STATE.uploadProgress
+      }
+      break
     case "COMPOSER.SET_DRAFT_STATUS":
       state = {
         ...state,
