+++ src/store.js
@@ -10,7 +10,6 @@ import list from "./reducers/listReducer"
 import article from "./reducers/articleReducer"
 import pictures from "./reducers/pictureReducer"
 import user from "./reducers/userReducer"
-import upload from "./reducers/uploadReducer"
 
 export default createStore(
   combineReducers({
@@ -20,8 +19,7 @@ export default createStore(
     list,
     article,
     pictures,
-    user,
-    upload
+    user
   }),
   window.__REDUX_DEVTOOLS_EXTENSION__ && window.__REDUX_DEVTOOLS_EXTENSION__(),
   applyMiddleware(/*logger(), promise(),*/ thunk)
